RAMEN: Regional Association of Methylome variability with the Exposome and geNome

RAMEN is an R package which goal is to identify Variable Methylated Loci (VML) in microarray DNA methylation data and then, using genome and exposome data, identify which individual genetic (G) or environmental (E) or joint (G+E or GxE) model better explains the VML variability. This package provides estimates of genetic and environmental contribution for each VML, allowing researchers to gain insights into the complex interplay between genetics, environment and DNA methylation.

See also

Author

Maintainer: Erick I. Navarro-Delgado erick.delgado@ubc.ca (ORCID)

Authors:

Erick I. Navarro-Delgado erick.delgado@ubc.ca (ORCID)

Keegan Korthauer keegan@stat.ubc.ca (ORCID)

Michael S. Kobor michael.kobor@ubc.ca (ORCID)

Other contributors:

Lluís Revilla Sancho lluis.revilla@gmail.com (ORCID) [reviewer]

Ulduz Vafadarshamasbi vafadar.u@uq.edu.au (ORCID) [reviewer]

The University of British Columbia [funder]